Use AdBlockingStatusChecker in AdBlockingExtensionJsInjectorPlugin
