+++ ad-blocking/ad-blocking-impl/src/main/java/com/duckduckgo/adblocking/impl/domain/AdBlockingStatusChecker.kt
@@ -16,10 +16,16 @@
 
 package com.duckduckgo.adblocking.impl.domain
 
+import com.duckduckgo.adblocking.impl.AdBlockingSettingsRepository
 import com.duckduckgo.adblocking.impl.remoteconfig.AdBlockingExtensionFeature
+import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesBinding
 import dagger.SingleInstanceIn
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.stateIn
 import logcat.logcat
 import javax.inject.Inject
 
@@ -31,7 +37,12 @@ interface AdBlockingStatusChecker {
 @ContributesBinding(AppScope::class)
 class RealAdBlockingStatusChecker @Inject constructor(
     private val feature: AdBlockingExtensionFeature,
+    settingsRepository: AdBlockingSettingsRepository,
+    @AppCoroutineScope appScope: CoroutineScope,
 ) : AdBlockingStatusChecker {
+
+    private val userEnabled: StateFlow<Boolean> = settingsRepository.isEnabledFlow()
+        .stateIn(appScope, SharingStarted.Eagerly, initialValue = false)
 
     override fun canInject(): Boolean {
         if (!feature.self().isEnabled()) {
@@ -40,6 +51,10 @@ class RealAdBlockingStatusChecker @Inject constructor(
         }
         if (feature.enableContingencyMode().isEnabled()) {
             logcat { "Contingency mode is on" }
+            return false
+        }
+        if (!userEnabled.value) {
+            logcat { "User disabled ad blocking" }
             return false
         }
         return true
