+++ ad-blocking/ad-blocking-impl/src/test/java/com/duckduckgo/adblocking/impl/domain/RealAdBlockingStatusCheckerTest.kt
@@ -16,8 +16,16 @@
 
 package com.duckduckgo.adblocking.impl.domain
 
+import com.duckduckgo.adblocking.impl.AdBlockingSettingsRepository
 import com.duckduckgo.adblocking.impl.remoteconfig.AdBlockingExtensionFeature
 import com.duckduckgo.feature.toggles.api.Toggle
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import org.junit.After
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
@@ -25,6 +33,7 @@ import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 
+@OptIn(ExperimentalCoroutinesApi::class)
 class RealAdBlockingStatusCheckerTest {
 
     private var killSwitchEnabled = true
@@ -41,7 +50,23 @@ class RealAdBlockingStatusCheckerTest {
         on { enableContingencyMode() } doReturn contingencyModeToggle
     }
 
-    private val checker = RealAdBlockingStatusChecker(feature)
+    private val userEnabledFlow = MutableStateFlow(true)
+    private val settingsRepository: AdBlockingSettingsRepository = object : AdBlockingSettingsRepository {
+        override fun isEnabledFlow(): Flow<Boolean> = userEnabledFlow
+        override suspend fun setEnabled(enabled: Boolean) {
+            userEnabledFlow.value = enabled
+        }
+    }
+    private val testScope = CoroutineScope(UnconfinedTestDispatcher())
+
+    private val checker by lazy {
+        RealAdBlockingStatusChecker(feature, settingsRepository, testScope)
+    }
+
+    @After
+    fun tearDown() {
+        testScope.cancel()
+    }
 
     @Test
     fun whenKillSwitchIsOnAndContingencyModeIsOffThenCanInject() {
@@ -58,6 +83,13 @@ class RealAdBlockingStatusCheckerTest {
     @Test
     fun whenContingencyModeIsOnThenCannotInject() {
         contingencyModeEnabled = true
+
+        assertFalse(checker.canInject())
+    }
+
+    @Test
+    fun whenUserHasDisabledThenCannotInject() {
+        userEnabledFlow.value = false
 
         assertFalse(checker.canInject())
     }
